+++ src/gpu/gpu_vertex_buffer.c
@@ -25,13 +25,13 @@ typedef struct {
 static GPUVertBuffer vert_buffers[VERT_BUFFER_MAX_COUNT];
 static short next_deleted_vert_buffer;
 
-static inline int find_deleted_vert_buffer_id(void)
+static inline short find_deleted_vert_buffer_id(void)
 {
   return ARRAY_FIND_DELETED_ID(next_deleted_vert_buffer, vert_buffers,
                             VERT_BUFFER_MAX_COUNT, GPUVertBuffer, "vertex buffer");
 }
 
-static GPUVertBuffer *get_vert_buffer(int id)
+static GPUVertBuffer *get_vert_buffer(short id)
 {
   if ((id < 0) | (id >= VERT_BUFFER_MAX_COUNT)) {
     id = 0;
@@ -69,9 +69,9 @@ static void vertbuf_init(GPUVertBuffer *vbuff)
 }
 
 // create new GPUVertBuffer
-int GPU_vertbuf_create(void)
+short GPU_vertbuf_create(void)
 {
-  const int id = find_deleted_vert_buffer_id();
+  const short id = find_deleted_vert_buffer_id();
   GPUVertBuffer *const vbuff = get_vert_buffer(id);
   vbuff->active = 1;
   vertbuf_init(vbuff);
@@ -90,17 +90,17 @@ void GPU_vertbuf_delete(const int id)
     next_deleted_vert_buffer = id;
 }
 
-void GPU_vertbuf_set_vertex_format(const int id, const int vformat)
+void GPU_vertbuf_set_vertex_format(const short id, const int vformat)
 {
   get_vert_buffer(id)->vformat = vformat;
 }
 
-void GPU_vertbuf_set_add_count(const int id, const GLuint count)
+void GPU_vertbuf_set_add_count(const short id, const GLuint count)
 {
   get_vert_buffer(id)->add_count = count;
 }
 
-void GPU_vertbuf_set_start(const int id, const GLuint start)
+void GPU_vertbuf_set_start(const short id, const GLuint start)
 {
   GPUVertBuffer *const vbuff = get_vert_buffer(id);
   
@@ -168,7 +168,7 @@ static void *vertbuf_attr_data(GPUVertBuffer *vbuff, const GLuint attribute_id)
   return vbuff->data + (idx * stride) + GPU_vertex_format_offset(vbuff->vformat, attribute_id);  
 }
 
-void GPU_vertbuf_add_4(const int id, const GLuint attribute_id, const GLfloat val1, const GLfloat val2, const GLfloat val3, const GLfloat val4)
+void GPU_vertbuf_add_4(const short id, const GLuint attribute_id, const GLfloat val1, const GLfloat val2, const GLfloat val3, const GLfloat val4)
 {
   GPUVertBuffer *const vbuff = get_vert_buffer(id);
 
@@ -186,7 +186,7 @@ void GPU_vertbuf_add_4(const int id, const GLuint attribute_id, const GLfloat va
   }
 }
 
-void GPU_vertbuf_use_BO(const int id)
+void GPU_vertbuf_use_BO(const short id)
 {
   GPUVertBuffer *const vbuff = get_vert_buffer(id);
 
@@ -195,13 +195,13 @@ void GPU_vertbuf_use_BO(const int id)
   }
 }
 
-void GPU_vertbuf_no_BO(const int id)
+void GPU_vertbuf_no_BO(const short id)
 {
   GPUVertBuffer *const vbuff = get_vert_buffer(id);
   delete_vbo(vbuff);
 }
 
-void GPU_vertbuf_bind(const int id)
+void GPU_vertbuf_bind(const short id)
 {
   GPUVertBuffer *const vbuff = get_vert_buffer(id);
